Add URI property to BonusSource
At the moment, it doesn't do anything with the property, but it's there.
Also, I wonder if it's worthwhile collecting all the objects with URIs
into a base class (maybe called Resource).

In any case, I'll duplicate it for now and change it if any more things
end up needing a URI.